requirements.txt
flask>=2.0.1
opencv-contrib-python>=4.5.5
numpy>=1.24.0
Pillow>=10.0.0
werkzeug>=2.3.7 